Address MDS review blockers for protocol and minting.
Map ActionController::BadRequest to plain-text MDS 400 responses so
handle/url invariant failures do not 500. Extract DoiMinting from
Helpable and use it from Mds::DoiMinter without the full model concern.
Fix REST head :no_content early returns, nil-safe get_dois client
lookup, drop unused render_mds 204 branch, and tighten MDS GET /doi
specs to deterministic stored-URL cases.
